+++ src/pygrowthstandards/oop/calculator.py
@@ -4,6 +4,7 @@ import os
 import pandas as pd
 
 from pygrowthstandards.data.transform import GrowthData
+from pygrowthstandards.utils.config import ChoiceValidator, MeasurementTypeType
 
 from ..utils import stats
 from ..utils.errors import NoReferenceDataException
@@ -17,21 +18,16 @@ class Calculator:
 
     path = "data"
 
-    x_var_types = {
-        "very_preterm_newborn": "gestational_age",
-        "newborn": "gestational_age",
-        "very_preterm_growth": "gestational_age",
-        "child_growth": "age",
-        "growth": "age",
-    }
-
     def __init__(self):
         self.data = pd.read_parquet(
             os.path.join(self.path, f"pygrowthstandards_{GrowthData.version}.parquet")
         )
 
     def calculate_z_score(
-        self, measurement_group: MeasurementGroup, measurement_type: str, age_value: int
+        self,
+        measurement_group: MeasurementGroup,
+        measurement_type: MeasurementTypeType,
+        age_value: int,
     ) -> float:
         value = getattr(measurement_group, measurement_type, None)
         if value is None:
@@ -39,11 +35,20 @@ class Calculator:
                 f"MeasurementGroup with age {age_value} does not have data for '{measurement_type}'."
             )
 
-        age_type = self.x_var_types[measurement_group.table_name]
+        assert measurement_group.age_group is not None, (
+            "MeasurementGroup must have an age_group to calculate z-score."
+        )
+        age_type = ChoiceValidator.get_age_type_from_age_group(
+            measurement_group.age_group
+        )
+        assert age_type is not None, (
+            f"No valid age type found for age group '{measurement_group.age_group}'."
+        )
 
         filtered_data = self._filter_measurement_data(
             self.data, measurement_type, age_type, age_value
         )
+
         L, M, S = self._get_lms_params(filtered_data, age_value)
 
         return stats.calculate_z_score(value, L, M, S)
@@ -53,14 +58,13 @@ class Calculator:
         measurement_group: MeasurementGroup,
         age_value: int,
     ) -> MeasurementGroup:
-        z_score_group = MeasurementGroup(
-            table_name=measurement_group.table_name, date=measurement_group.date
-        )
+        z_score_group = MeasurementGroup(date=measurement_group.date)
+        z_score_group.age_group = measurement_group.age_group
 
         data = measurement_group.to_dict()
 
         for key, value in data.items():
-            if value is None or key in ["date", "table_name"]:
+            if value is None or key in ["date", "age_group"]:
                 continue
 
             try: